+++ src/lib/pipe.ts
@@ -1,53 +1,15 @@
-// Delta function -- first in composition can have any number of inputs, single output
-export type DeltaFn0<TRes> = () => TRes | Promise<TRes> | TRes[] | Promise<TRes[]>;
-type DeltaFn1<T1, TRes> = (p1: T1) => TRes | Promise<TRes> | TRes[] | Promise<TRes[]>;
-type DeltaFn2<T1, T2, TRes> = (p1: T1, p2: T2) => TRes | Promise<TRes> | TRes[] | Promise<TRes[]>;
-type DeltaFn3<T1, T2, T3, TRes> = (p1: T1, p2: T2, p3: T3) => TRes | Promise<TRes> | TRes[] | Promise<TRes[]>;
-type DeltaFn4<T1, T2, T3, T4, TRes> = (p1: T1, p2: T2, p3: T3, p4: T4) => TRes | Promise<TRes> | TRes[] | Promise<TRes[]>;
-type DeltaFn5<T1, T2, T3, T4, T5, TRes> = (p1: T1, p2: T2, p3: T3, p4: T4, p5: T5) => TRes | Promise<TRes> | TRes[] | Promise<TRes[]>;
-export type DeltaFn<T1, T2, T3, T4, T5, TRes> =
-    DeltaFn0<TRes>
-    | DeltaFn1<T1, TRes>
-    | DeltaFn2<T1, T2, TRes>
-    | DeltaFn3<T1, T2, T3, TRes>
-    | DeltaFn4<T1, T2, T3, T4, TRes>
-    | DeltaFn5<T1, T2, T3, T4, T5, TRes>;
-
-// River function -- other functions in composition flow: one input, one output
-type RiverFn<T, TRes> = ((p: T) => TRes | Promise<TRes>) | ((p: T[]) => TRes | Promise<TRes>);
-
-// Shorthand for return type of pipe (only replace return value of Delta function)
-// see https://stackoverflow.com/a/50014868
-type ArgumentTypes<T> = T extends (...args: infer U) => infer R ? U : never;
-type ReplaceReturnTypePromise<T, TNewReturn> = (...a: ArgumentTypes<T>) => Promise<TNewReturn>;
-
-// Pipe (0-5 functions, first Delta)
+import { ArgType, LastIndexOf, Lookup, ReplaceReturnTypePromise, Tail, UnpackPromise, VariadicFunction } from './internal/metaTypes';
+type RiverFn = (arg: any) => any;
+type AsChain<F extends [RiverFn, ...RiverFn[]], G extends RiverFn[]= Tail<F>> = {
+    [K in keyof F]: (arg: UnpackPromise<ArgType<F[K]>>) => ArgType<Lookup<G, K, any>, any>
+};
 function pipe<T>(): (arg: T) => Promise<T>;
-function pipe<D1, D2, D3, D4, D5, DRes>(df: DeltaFn<D1, D2, D3, D4, D5, DRes>):
-    ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, DRes>;
-function pipe<D1, D2, D3, D4, D5, DRes, R1>(df: DeltaFn<D1, D2, D3, D4, D5, DRes>, r1: RiverFn<DRes, R1>):
-    ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, R1>;
-function pipe<D1, D2, D3, D4, D5, DRes, R1, R2>(df: DeltaFn<D1, D2, D3, D4, D5, DRes>, r1: RiverFn<DRes, R1>, r2: RiverFn<R1, R2>):
-    ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, R2>;
-function pipe<D1, D2, D3, D4, D5, DRes, R1, R2, R3>(df: DeltaFn<D1, D2, D3, D4, D5, DRes>, r1: RiverFn<DRes, R1>, r2: RiverFn<R1, R2>, r3: RiverFn<R2, R3>):
-    ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, R3>;
-function pipe<D1, D2, D3, D4, D5, DRes, R1, R2, R3, R4>(
-    df: DeltaFn<D1, D2, D3, D4, D5, DRes>,
-    r1: RiverFn<DRes, R1>,
-    r2: RiverFn<R1, R2>,
-    r3: RiverFn<R2, R3>,
-    r4: RiverFn<R3, R4>
-): ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, R4>;
-function pipe<D1, D2, D3, D4, D5, DRes, R1, R2, R3, R4, R5>(
-    df: DeltaFn<D1, D2, D3, D4, D5, DRes>,
-    r1: RiverFn<DRes, R1>,
-    r2: RiverFn<R1, R2>,
-    r3: RiverFn<R2, R3>,
-    r4: RiverFn<R3, R4>,
-    r5: RiverFn<R4, R5>
-): ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, R5>;
-function pipe<D1, D2, D3, D4, D5, DRes, Res>(df: DeltaFn<D1, D2, D3, D4, D5, DRes>, ...fns: any[]):
-    ReplaceReturnTypePromise<DeltaFn<D1, D2, D3, D4, D5, DRes>, any>;
+function pipe<Delta extends VariadicFunction>(df: Delta): Delta;
+function pipe<
+    Delta extends VariadicFunction,
+    F extends [(arg: UnpackPromise<ReturnType<Delta>>) => any, ...Array<(arg: any) => any>]
+>(df: Delta, ...rivers: F & AsChain<F>):
+    ReplaceReturnTypePromise<Delta, ReturnType<F[LastIndexOf<F>]>>;
 
 /**
  * Create a function composed of provided functions in left-to-right execution chain.
@@ -83,7 +45,7 @@ function pipe<D1, D2, D3, D4, D5, DRes, Res>(df: DeltaFn<D1, D2, D3, D4, D5, DRe
  * @return Function with signature of the first function (or no args if no functions provided) returning Promise of a result of last function
  */
 function pipe(...fns: any[]): any {
-    return (...initialArgs: any) =>
+    return (...initialArgs: any[]) =>
         fns.reduce((pendingLastResult, fn, i) => pendingLastResult.then((lastResult: any) => {
             // first iteration (lastResult is initialArgs), spread into delta function
             const currentResult = i === 0 ? fn(...lastResult) : fn(lastResult);