+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2018-12-10
 ### Added
 - Pipe to readme
 
@@ -94,7 +96,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `conf.json`
 - `jsconfig.json`
 
-[Unreleased]: https://github.com/AckeeCZ/desmond/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/AckeeCZ/desmond/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/AckeeCZ/desmond/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/AckeeCZ/desmond/compare/v0.2.5...v0.3.0
 [0.2.5]: https://github.com/AckeeCZ/desmond/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/AckeeCZ/desmond/compare/v0.2.3...v0.2.4